refine std_base64_decode function.

+++ src/client/client.c
@@ -1204,7 +1204,7 @@ static void tls_cli_on_connection_established(struct tls_cli_ctx* tls_cli, int s
             buf = websocket_connect_request(domain, domain_port, url_path, ctx->sec_websocket_key, &malloc, &len);
             if (config->target_address)
             {
-                char* b64addr = url_safe_base64_encode_alloc(typ, (int)typ_len, &malloc);
+                char* b64addr = url_safe_base64_encode_alloc(typ, (size_t)typ_len, &malloc);
                 static const char* addr_fmt = "Target-Address" ": %s\r\n";
                 char* addr_field = (char*)calloc(strlen(addr_fmt) + strlen(b64addr) + 1, sizeof(*addr_field));
                 sprintf(addr_field, addr_fmt, b64addr);
@@ -1218,7 +1218,7 @@ static void tls_cli_on_connection_established(struct tls_cli_ctx* tls_cli, int s
             if (ctx->udp_data_ctx) {
                 size_t addr_len = 0;
                 uint8_t* addr_p = socks5_address_binary(&ctx->udp_data_ctx->target_addr, &malloc, &addr_len);
-                char* b64str = url_safe_base64_encode_alloc(addr_p, (int)addr_len, &malloc);
+                char* b64str = url_safe_base64_encode_alloc(addr_p, (size_t)addr_len, &malloc);
                 static const char* udp_fmt = "UDP" ": %s\r\n";
                 char* udp_field = (char*)calloc(strlen(udp_fmt) + strlen(b64str) + 1, sizeof(*udp_field));
                 sprintf(udp_field, udp_fmt, b64str);
